Coloquei o form do editProfile em um componente
